// prisma/generated/internal/prismaNamespace.ts
export const QueryMode = {
    default: "default",
    insensitive: "insensitive",
} as const

export type QueryMode = (typeof QueryMode)[keyof typeof QueryMode]